move files to groups directory
