+++ src/groups/GroupDetails.tsx
@@ -1,15 +1,15 @@
 import { SpinnerOverlay, useAutoLoadedData } from '@terra-ui-packages/components';
 import _ from 'lodash/fp';
 import React, { useEffect, useState } from 'react';
-import { DeleteUserModal } from 'src/components/DeleteUserModal';
-import { EditUserModal } from 'src/components/EditUserModal';
 import FooterWrapper from 'src/components/FooterWrapper';
 import { DelayedSearchInput } from 'src/components/input';
-import { Member, MemberTable } from 'src/components/MemberTable';
-import { NewUserModal } from 'src/components/NewUserModal';
 import { PageBox, PageBoxVariants } from 'src/components/PageBox';
 import { TopBar } from 'src/components/TopBar';
 import { AdminNotifierCheckbox } from 'src/groups/AdminNotifierCheckbox';
+import { DeleteUserModal } from 'src/groups/Members/DeleteUserModal';
+import { EditUserModal } from 'src/groups/Members/EditUserModal';
+import { Member, MemberTable } from 'src/groups/Members/MemberTable';
+import { NewUserModal } from 'src/groups/Members/NewUserModal';
 import { Ajax } from 'src/libs/ajax';
 import { GroupRole } from 'src/libs/ajax/Groups';
 import { reportError } from 'src/libs/error';
